# PERF: Do not init cache in RangeIndex.take