--- a/pandas/tests/indexes/ranges/test_range.py
+++ b/pandas/tests/indexes/ranges/test_range.py
@@ -210,7 +210,7 @@ class TestRangeIndex:
         assert index.dtype == np.int64
 
     def test_cache(self):
-        # GH 26565, GH26617, GH35432
+        # GH 26565, GH26617, GH35432, GH53387
         # This test checks whether _cache has been set.
         # Calling RangeIndex._cache["_data"] creates an int64 array of the same length
         # as the RangeIndex and stores it in _cache.
@@ -264,23 +264,7 @@ class TestRangeIndex:
         df.iloc[5:10]
         assert idx._cache == {}
 
-        # idx._cache should contain a _data entry after call to idx._data
-        idx._data
-        assert isinstance(idx._data, np.ndarray)
-        assert idx._data is idx._data  # check cached value is reused
-        assert len(idx._cache) == 1
-        expected = np.arange(0, 100, 10, dtype="int64")
-        tm.assert_numpy_array_equal(idx._cache["_data"], expected)
-
-    def test_cache_after_calling_loc_with_array(self):
-        # GH 53387
-        # the cache will contain a _constructor key, so it should be tested separately
-        idx = RangeIndex(0, 100, 10)
-        df = pd.DataFrame({"a": range(10)}, index=idx)
-
-        assert "_data" not in idx._cache
-
-        # take is internally called by loc, but it's also tested explicitly
+        # after calling take, _cache may contain other keys, but not "_data"
         idx.take([3, 0, 1])
         assert "_data" not in idx._cache
 
@@ -289,6 +273,14 @@ class TestRangeIndex:
 
         df.iloc[[5, 6, 7, 8, 9]]
         assert "_data" not in idx._cache
+
+        # idx._cache should contain a _data entry after call to idx._data
+        idx._data
+        assert isinstance(idx._data, np.ndarray)
+        assert idx._data is idx._data  # check cached value is reused
+        assert "_data" in idx._cache
+        expected = np.arange(0, 100, 10, dtype="int64")
+        tm.assert_numpy_array_equal(idx._cache["_data"], expected)
 
     def test_is_monotonic(self):
         index = RangeIndex(0, 20, 2)