@GianlucaFicarelli 2023-07-18
> Couple questions, otherwise LGTM

Thanks, I've answered all the pending questions.